feature: adição de novos módulos

+++ src/modules/countries/controllers/country.controller.ts
@@ -15,20 +15,17 @@ import { CreateCountryDto } from '../dto/create-country.dto';
 import { CountryService } from '../services/country.service';
 import { ApiOperation, ApiTags } from '@nestjs/swagger';
 import { CountryEntity } from '../entities/country.entity';
-import { ApiResponses } from '../../../utils/decorators';
 import { UpdateCountryDto } from '../dto/update-country.dto';
 import { isNumber } from '@nestjs/class-validator';
 import { FilterCountryDto } from '../dto/filter-country.dto';
-
+import { countryDocumentation } from '../documentation';
+const { ApiOperation: doc } = countryDocumentation;
 @ApiTags('countries')
 @Controller('country')
 export class CountryController {
   constructor(private countryService: CountryService) {}
-  @ApiOperation({
-    summary: 'country/:id',
-    description:
-      'Este endpoint recebe como param o id e retorna os dados do país.',
-  })
+
+  @ApiOperation(doc.getById)
   @Get('getById/:id')
   @UsePipes(new ValidationPipe())
   async getById(@Param('id') id: number): Promise<CountryEntity> {
@@ -38,13 +35,8 @@ export class CountryController {
     return await this.countryService.findById(id);
   }
 
-  @ApiOperation({
-    summary: 'country/create',
-    description:
-      'Este endpoint recebe como body o name e language para salvar um registro de dados.',
-  })
+  @ApiOperation(doc.createCountry)
   @Post('create')
-  @ApiResponses({ 201: CreateCountryDto })
   @UsePipes(new ValidationPipe())
   async createCountry(
     @Body() newCountry: CreateCountryDto,
@@ -52,16 +44,16 @@ export class CountryController {
     return this.countryService.createCountry(newCountry);
   }
 
-  @ApiOperation({
-    summary: 'country/update/:id',
-    description:
-      'Este endpoint recebe como body o name e language e o path como id para atualizar um registro de dados.',
-  })
+  @ApiOperation(doc.updateCountry)
   @Patch('update/:id')
+  @UsePipes(new ValidationPipe())
   async updateCountry(
     @Param('id') id: number,
     @Body() updateCountryDto: UpdateCountryDto,
   ): Promise<CountryEntity> {
+    if (!isNumber(id)) {
+      throw new BadRequestException('FieldMustBeNumber');
+    }
     const countryUpdate = await this.countryService.updateCountry(
       id,
       updateCountryDto,
@@ -69,15 +61,17 @@ export class CountryController {
     return countryUpdate;
   }
 
-  @ApiOperation({
-    summary: 'country/:id',
-    description: 'Este endpoint recebe como param o id e excluí o registro',
-  })
+  @ApiOperation(doc.deleteById)
   @Delete(':id')
+  @UsePipes(new ValidationPipe())
   async deleteById(@Param('id') id: number): Promise<string> {
+    if (!isNumber(id)) {
+      throw new BadRequestException('FieldMustBeNumber');
+    }
     return await this.countryService.deleteCountry(id);
   }
 
+  @ApiOperation(doc.getByFilter)
   @Get('getByFilter')
   @UsePipes(new ValidationPipe())
   async getByFilter(
